+++ CHANGELOG.md
@@ -21,7 +21,8 @@ This release adds support for 19-digit cards in `CardInputWidget` and `CardMulti
     * Upgrade `com.nimbusds:nimbus-jose-jwt` to `9.0.1`
     * Guard against crash when `TransactionTimer` is unavailable
 * [#2873](https://github.com/stripe/stripe-android/pull/2873) Fix `CardInputWidget` field rendering in RTL
-* [#2873](https://github.com/stripe/stripe-android/pull/2878) Remove `Stripe.createToken()`
+* [#2878](https://github.com/stripe/stripe-android/pull/2878) Remove `Stripe.createToken()`
+* [#2880](https://github.com/stripe/stripe-android/pull/2880) Fix date formatting in `KlarnaSourceParams`
 
 ## 15.1.0 - 2020-08-13
 * [#2671](https://github.com/stripe/stripe-android/pull/2671) Add `cardParams` property to `CardInputWidget` and `CardMultilineWidget`